+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - A user's API key can be attained without a super user token.
 - More specialised errors.
 - Change to name of verify flags on User ORM's create and delete methods to unify them.
+- Method to modify user now returns whether a modification occurs, without requiring a modified Shinobi installation.
 
 ### Added
 - Monitor ORM.